+++ rad_tools/__init__.py
@@ -2,5 +2,5 @@
 All The tools from the package.
 """
 
-__version__ = "0.0.0.5"
+__version__ = "0.0.0.6"